Rmoved odd using statement from ICommitmentsService, modified unit tests to check the view is mapped correctly and to more efficiently check if AnyApprenticeships is set, and added a null check for when assigning to AnyApprenticeships in the model and associated unit test

+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Controllers/ManageApprenticesController.cs
@@ -17,7 +17,7 @@ namespace SFA.DAS.ProviderCommitments.Web.Controllers
         }
 
         [Route("all")]
-        public async Task<IActionResult> Index(long providerId)
+        public async Task<IActionResult> Index(uint providerId)
         {
             if (!ModelState.IsValid)
             {
@@ -27,7 +27,7 @@ namespace SFA.DAS.ProviderCommitments.Web.Controllers
             var model = new ManageApprenticesViewModel
             {
                 ProviderId = providerId,
-                Apprenticeships = await _commitmentsService.GetApprovedApprenticeships((uint)providerId)
+                Apprenticeships = await _commitmentsService.GetApprovedApprenticeships(providerId)
             };
 
             return View(model);
